fix: yMax = NaN
Since the API sometimes returns JSON values with not all fields set
(that's a bug which is being worked around here), it can happen that
the size-field is 0. This results in a division-by-0 when calculating
the feerate and the max value for the y-axis can't be calculated. No
transactions are drawn. This is fixed here by filtering the JSON
values where not all fields are set.

I didn't futher investigate why the API returns incorrect JSON values.